add rate values component
