+++ src/components/offer/offer/review-form/review-form.tsx
@@ -1,9 +1,11 @@
 import { useAppDispatch, useAppSelector } from '@/hooks';
-import { ReviewLength } from '@/libs/const';
+import { RateValues, RequestStatus, ReviewLength } from '@/libs/const';
 import { ReviewToSend } from '@/libs/types/types';
 import { offerFullSElectors } from '@/storage/slices/fullOffer';
 import { sendReview } from '@/thunk/fullOffer';
 import { ChangeEvent, useState } from 'react';
+import { ReviewRate } from '../../reviews/review-rate/review-rate';
+
 
 function ReviewForm() {
 
@@ -12,6 +14,8 @@ function ReviewForm() {
   const currentOffer = useAppSelector(offerFullSElectors.selectFullOffer);
 
   const isDisable = useAppSelector(offerFullSElectors.selectIsFormDisabled);
+
+  const postStatus = useAppSelector(offerFullSElectors.selectPostReviewRequestStatus);
 
   const [reviewText, setReviewText] = useState('');
 
@@ -32,6 +36,10 @@ function ReviewForm() {
       const id = currentOffer.id;
       dispatch(sendReview({id, review}));
     }
+
+    if (postStatus === RequestStatus.Success) {
+      setReviewText('');
+    }
   };
 
   const handleStarClick = (e: React.MouseEvent<HTMLInputElement>) => {
@@ -43,38 +51,9 @@ function ReviewForm() {
     <form onSubmit={handleSubmitForm} className="reviews__form form" action="#" method="post">
       <label className="reviews__label form__label" htmlFor="review">Your review</label>
       <div className="reviews__rating-form form__rating">
-        <input className="form__rating-input visually-hidden" name="rating" defaultValue={5} id="5-stars" type="radio" onClick={handleStarClick} />
-        <label htmlFor="5-stars" className="reviews__rating-label form__rating-label" title="perfect">
-          <svg className="form__star-image" width={37} height={33}>
-            <use xlinkHref="#icon-star" />
-          </svg>
-        </label>
-        <input className="form__rating-input visually-hidden" name="rating" defaultValue={4} id="4-stars" type="radio" onClick={handleStarClick}/>
-        <label htmlFor="4-stars" className="reviews__rating-label form__rating-label" title="good">
-          <svg className="form__star-image" width={37} height={33}>
-            <use xlinkHref="#icon-star" />
-          </svg>
-        </label>
-        <input className="form__rating-input visually-hidden" name="rating" defaultValue={3} id="3-stars" type="radio" onClick={handleStarClick}/>
-        <label htmlFor="3-stars" className="reviews__rating-label form__rating-label" title="not bad">
-          <svg className="form__star-image" width={37} height={33}>
-            <use xlinkHref="#icon-star" />
-          </svg>
-        </label>
-        <input className="form__rating-input visually-hidden" name="rating" defaultValue={2} id="2-stars" type="radio" onClick={handleStarClick}/>
-        <label htmlFor="2-stars" className="reviews__rating-label form__rating-label" title="badly">
-          <svg className="form__star-image" width={37} height={33}>
-            <use xlinkHref="#icon-star" />
-          </svg>
-        </label>
-        <input className="form__rating-input visually-hidden" name="rating" defaultValue={1} id="1-star" type="radio" onClick={handleStarClick}/>
-        <label htmlFor="1-star" className="reviews__rating-label form__rating-label" title="terribly">
-          <svg className="form__star-image" width={37} height={33}>
-            <use xlinkHref="#icon-star" />
-          </svg>
-        </label>
+        {Object.entries(RateValues).map(([rating, title]) => <ReviewRate key={title} isDisable={isDisable} handleStarClick={handleStarClick} defaultValue={Number(rating)} valueDescription={title}/>)}
       </div>
-      <textarea className="reviews__textarea form__textarea" name="review" minLength={ReviewLength.Min} maxLength={ReviewLength.Max} value={reviewText} id="review" placeholder="Tell how was your stay, what you like and what can be improved" onChange={handleTextChange}/>
+      <textarea className="reviews__textarea form__textarea" name="review" minLength={ReviewLength.Min} maxLength={ReviewLength.Max} value={reviewText} id="review" placeholder="Tell how was your stay, what you like and what can be improved" required onChange={handleTextChange}/>
       <div className="reviews__button-wrapper">
         <p className="reviews__help">
           To submit review please make sure to set <span className="reviews__star">rating</span> and describe your stay with at least <b className="reviews__text-amount">{ReviewLength.Min} characters</b>.
