+++ src/pages/main-page/main-page.tsx
@@ -2,8 +2,8 @@ import './main-page.css';
 import MainCitiesLocations from '@/components/main/main-cities-locations/main-cities-locations';
 import MainCitiesContainer from '@/components/main/main-cities-container/main-cities-container';
 import { CITIES_LOCATIONS, DEFAULT_CITY } from '@/libs/mocks/cities-locations';
-import { Cities } from '@/libs/const';
-import { OfferCardPrew, RequestStatus } from '@/libs/types/types';
+import { Cities, RequestStatus } from '@/libs/const';
+import { OfferCardPrew } from '@/libs/types/types';
 import Spinner from '@/components/spinner/spinner';
 import MainEmpty from '../main-empty/main-empty';
 import { offersSelectors } from '@/storage/slices/offers';
