+++ src/storage/slices/offers.ts
@@ -1,7 +1,7 @@
 import { createSlice, type PayloadAction } from '@reduxjs/toolkit';
 import { OFFERS_SLICE_NAME } from './sliceNames';
-import { Cities, SortItem } from '@/libs/const';
-import { OfferCardPrew, RequestStatus } from '@/libs/types/types';
+import { Cities, RequestStatus, SortItem } from '@/libs/const';
+import { OfferCardPrew } from '@/libs/types/types';
 import { fetchOffers } from '@/thunk/offers';
